fix: move tests to presentation project
